Only update communities after cache expires

+++ communities/src/main/java/com/neaniesoft/vermilion/communities/data/database/CommunityRepository.kt
@@ -13,6 +13,7 @@ import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import java.time.Clock
+import java.time.Duration
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -28,21 +29,30 @@ class CommunityRepositoryImpl @Inject constructor(
     private val api: CommunitiesApiService,
     private val clock: Clock
 ) : CommunityRepository {
+    companion object {
+        private val CACHE_TIMEOUT = Duration.ofDays(1).toMillis()
+    }
+
     override fun subscribedCommunities(): Flow<List<Community>> {
         return dao.observeAllSubscribedCommunities()
             .map { it.map { record -> record.toCommunity() } }
     }
 
     override suspend fun updateSubscribedCommunities() {
-        val communities =
-            getSubscribedCommunitiesFromApi(
-                null,
-                0
-            ) // Will loop through the pages until we've got them all
+        val lastInsertedTime = database.withTransaction { dao.getLastInsertedTime() } ?: 0
+        val shouldUpdate = clock.millis() > lastInsertedTime + CACHE_TIMEOUT
 
-        database.withTransaction {
-            dao.removeAllCommunities()
-            dao.insertAll(communities.map { it.toCommunityRecord(clock) })
+        if (shouldUpdate) {
+            val communities =
+                getSubscribedCommunitiesFromApi(
+                    null,
+                    0
+                ) // Will loop through the pages until we've got them all
+
+            database.withTransaction {
+                dao.removeAllCommunities()
+                dao.insertAll(communities.map { it.toCommunityRecord(clock) })
+            }
         }
     }
